+++ PropellantConfigBase.cs
@@ -4,9 +4,9 @@ namespace Ignition
 {
     public abstract class PropellantConfigBase
     {
-        public virtual float ThrustMultiplier { get; protected set; } = 1;
-        public virtual float IspMultiplier { get; protected set; } = 1;
-        public virtual float IgnitionPotential { get; protected set; } = 1;
+        public virtual double ThrustMultiplier { get; protected set; } = 1;
+        public virtual double IspMultiplier { get; protected set; } = 1;
+        public virtual double IgnitionPotential { get; protected set; } = 1;
         public List<Propellant> Propellants { get; protected set; } = new List<Propellant>();
 
         public PropellantConfigBase()
@@ -16,9 +16,9 @@ namespace Ignition
 
         public PropellantConfigBase(ConfigNode node)
         {
-            if (node.HasValue("ThrustMultiplier")) ThrustMultiplier = float.Parse(node.GetValue("ThrustMultiplier"));
-            if (node.HasValue("IspMultiplier")) IspMultiplier = float.Parse(node.GetValue("IspMultiplier"));
-            if (node.HasValue("IgnitionPotential")) IgnitionPotential = float.Parse(node.GetValue("IgnitionPotential"));
+            if (node.HasValue("ThrustMultiplier")) ThrustMultiplier = double.Parse(node.GetValue("ThrustMultiplier"));
+            if (node.HasValue("IspMultiplier")) IspMultiplier = double.Parse(node.GetValue("IspMultiplier"));
+            if (node.HasValue("IgnitionPotential")) IgnitionPotential = double.Parse(node.GetValue("IgnitionPotential"));
         }
     }
 }